Create Promised Singleton (#1)
A generic Typescript class for lazy-loading singletons. This is used in the donation form to lazy load the various payment clients from Braintree, PayPal, Apple, etc asynchronously without creating multiple instances of them.

PromisedSingleton is a generic wrapper for an asynchronously generated object that you only want one instance of (a singleton).